pyproject.toml
[tool.poetry]
name = "ofx2csv"
version = "0.1.0"
description = "Convert OFX files to CSV files"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ofxparse = "^0.21"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ofx2csv = "ofx2csv.ofx2csv:main"